Create README.md
An Python Program demonstrating the power of Reinforcement Learning, by making an AI Agent learn how to play the game of Snake. Using PyGame, a simple Snake game was made, and an agent was made that played the game. The Agent slowly learnt how to play the game through a reward-punishment system, which is the base of Reinforcement Learning. Deep Q-Learning was implemented with a PyTorch Neural Network, with the Bellman Equation being used as the Q-value predictor.